# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChittyOS GPT Connector API",
    "description": "Comprehensive API for connecting custom GPTs to ChittyOS services, third-party integrations, and banking connections",
    "version": "1.0.0",
    "contact": {
      "name": "ChittyOS Support",
      "url": "https://chitty.cc"
    }
  },
  "servers": [
    {
      "url": "https://connect.chitty.cc",
      "description": "Production ChittyConnect API"
    },
    {
      "url": "https://connect-staging.chitty.cc",
      "description": "Staging environment"
    }
  ],
  "security": [
    {
      "apiKey": []
    }
  ],
  "paths": {
...
    "/api/chittyid/mint": {
      "post": {
        "summary": "Mint ChittyID",
        "description": "Create a new ChittyID for an entity",
        "operationId": "mintChittyID",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "entity": {
                    "type": "string",
                    "enum": ["PEO", "PLACE", "PROP", "EVNT", "AUTH", "INFO", "FACT", "CONTEXT", "ACTOR"],
                    "description": "Entity type for ChittyID"
                  },
                  "metadata": {
                    "type": "object",
                    "description": "Optional metadata for the entity"
                  }
                },
                "required": ["entity"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ChittyID minted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChittyIDResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/chittyid/validate": {
      "post": {
        "summary": "Validate ChittyID",
        "description": "Validate a ChittyID format and checksum",
        "operationId": "validateChittyID",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "chittyid": {
                    "type": "string",
                    "description": "ChittyID to validate"
                  }
                },
                "required": ["chittyid"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": {
                      "type": "boolean"
                    },
                    "entity": {
                      "type": "string"
                    },
                    "sequence": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chittyfinance/account/balance": {
      "get": {
        "summary": "Get Account Balance",
        "description": "Get balance for ChittyFinance account",
        "operationId": "getAccountBalance",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Account balance retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string"
                    },
                    "balance": {
                      "type": "number"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "lastUpdated": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chittyfinance/banking/connect": {
      "post": {
        "summary": "Connect Banking Account",
        "description": "Connect external banking account via Plaid/Stripe",
        "operationId": "connectBankAccount",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "string",
                    "enum": ["plaid", "stripe", "direct"]
                  },
                  "publicToken": {
                    "type": "string",
                    "description": "Plaid public token or equivalent"
                  },
                  "accountDetails": {
                    "type": "object"
                  }
                },
                "required": ["provider"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bank account connected",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accountId": {
                      "type": "string"
                    },
                    "bankName": {
                      "type": "string"
                    },
                    "lastFour": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chittychronicle/log": {
      "post": {
        "summary": "Create Chronicle Entry",
        "description": "Log an event to ChittyChronicle",
        "operationId": "createChronicleEntry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "eventType": {
                    "type": "string"
                  },
                  "entityId": {
                    "type": "string"
                  },
                  "data": {
                    "type": "object"
                  },
                  "timestamp": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": ["eventType", "data"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry logged",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entryId": {
                      "type": "string"
                    },
                    "chronicleId": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chittychronicle/query": {
      "get": {
        "summary": "Query Chronicle",
        "description": "Query ChittyChronicle event logs",
        "operationId": "queryChronicle",
        "parameters": [
          {
            "name": "entityId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-ChittyOS-API-Key"
      }
    },
    "schemas": {
...
      "ChittyIDResponse": {
        "type": "object",
        "properties": {
          "chittyid": {
            "type": "string",
            "description": "Generated ChittyID"
          },
          "entity": {
            "type": "string"
          },
          "sequence": {
            "type": "string"
          },
          "checksum": {
            "type": "string"
          }
        }
      },
...
    }
  }
}